--- openmls/src/ciphersuite/ciphersuites.rs
@@ -1,181 +1,0 @@
-//! Ciphersuite and algorithm definitions and conversions.
-
-use openmls_traits::types::{HpkeAeadType, HpkeKdfType, HpkeKemType};
-
-use super::*;
-
-pub(crate) use std::convert::TryFrom;
-
-/// MLS ciphersuites.
-#[allow(non_camel_case_types)]
-#[allow(clippy::upper_case_acronyms)]
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    PartialEq,
-    Eq,
-    Hash,
-    Serialize,
-    Deserialize,
-    TlsDeserialize,
-    TlsSerialize,
-    TlsSize,
-)]
-#[repr(u16)]
-#[allow(missing_docs)]
-pub enum CiphersuiteName {
-    MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519 = 0x0001,
-    MLS10_128_DHKEMP256_AES128GCM_SHA256_P256 = 0x0002,
-    MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 = 0x0003,
-    MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448 = 0x0004,
-    MLS10_256_DHKEMP521_AES256GCM_SHA512_P521 = 0x0005,
-    MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 = 0x0006,
-    MLS10_256_DHKEMP384_AES256GCM_SHA384_P384 = 0x0007,
-}
-
-impl Default for CiphersuiteName {
-    /// The default ciphersuite is the MTI.
-    fn default() -> Self {
-        CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-    }
-}
-
-implement_enum_display!(CiphersuiteName);
-
-impl From<CiphersuiteName> for SignatureScheme {
-    #[inline(always)]
-    fn from(ciphersuite_name: CiphersuiteName) -> Self {
-        match ciphersuite_name {
-            CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519 => {
-                SignatureScheme::ED25519
-            }
-            CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256 => {
-                SignatureScheme::ECDSA_SECP256R1_SHA256
-            }
-            CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-                SignatureScheme::ED25519
-            }
-            CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448 => SignatureScheme::ED448,
-            CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521 => {
-                SignatureScheme::ECDSA_SECP521R1_SHA512
-            }
-            CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-                SignatureScheme::ED448
-            }
-            CiphersuiteName::MLS10_256_DHKEMP384_AES256GCM_SHA384_P384 => {
-                SignatureScheme::ECDSA_SECP384R1_SHA384
-            }
-        }
-    }
-}
-
-impl From<&CiphersuiteName> for u16 {
-    #[inline(always)]
-    fn from(s: &CiphersuiteName) -> u16 {
-        *s as u16
-    }
-}
-
-impl TryFrom<u16> for CiphersuiteName {
-    type Error = tls_codec::Error;
-
-    #[inline(always)]
-    fn try_from(v: u16) -> Result<Self, Self::Error> {
-        match v {
-            0x0001 => Ok(CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519),
-            0x0002 => Ok(CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256),
-            0x0003 => Ok(CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519),
-            0x0004 => Ok(CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448),
-            0x0005 => Ok(CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521),
-            0x0006 => Ok(CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448),
-            _ => Err(Self::Error::DecodingError(format!(
-                "{} is not a valid cipher suite value",
-                v
-            ))),
-        }
-    }
-}
-
-#[inline(always)]
-pub(crate) fn kem_from_suite(ciphersuite_name: &CiphersuiteName) -> HpkeKemType {
-    match ciphersuite_name {
-        CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-        | CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-            HpkeKemType::DhKem25519
-        }
-        CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256 => HpkeKemType::DhKemP256,
-        CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448
-        | CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-            HpkeKemType::DhKem448
-        }
-        CiphersuiteName::MLS10_256_DHKEMP384_AES256GCM_SHA384_P384 => HpkeKemType::DhKemP384,
-        CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521 => HpkeKemType::DhKemP521,
-    }
-}
-
-#[inline(always)]
-pub(crate) fn hpke_kdf_from_suite(ciphersuite_name: &CiphersuiteName) -> HpkeKdfType {
-    match ciphersuite_name {
-        CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-        | CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256
-        | CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-            HpkeKdfType::HkdfSha256
-        }
-        CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448
-        | CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521
-        | CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-            HpkeKdfType::HkdfSha512
-        }
-        CiphersuiteName::MLS10_256_DHKEMP384_AES256GCM_SHA384_P384 => HpkeKdfType::HkdfSha384,
-    }
-}
-
-#[inline(always)]
-pub(crate) fn hpke_aead_from_suite(ciphersuite_name: &CiphersuiteName) -> HpkeAeadType {
-    match ciphersuite_name {
-        CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-        | CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256 => HpkeAeadType::AesGcm128,
-        CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-            HpkeAeadType::ChaCha20Poly1305
-        }
-        CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448
-        | CiphersuiteName::MLS10_256_DHKEMP384_AES256GCM_SHA384_P384
-        | CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521 => HpkeAeadType::AesGcm256,
-        CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-            HpkeAeadType::ChaCha20Poly1305
-        }
-    }
-}
-
-#[inline(always)]
-pub(crate) fn hash_from_suite(ciphersuite_name: &CiphersuiteName) -> HashType {
-    match ciphersuite_name {
-        CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-        | CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256
-        | CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-            HashType::Sha2_256
-        }
-        CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448
-        | CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521
-        | CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => HashType::Sha2_512,
-        CiphersuiteName::MLS10_256_DHKEMP384_AES256GCM_SHA384_P384 => HashType::Sha2_384,
-    }
-}
-
-#[inline(always)]
-pub(crate) fn aead_from_suite(ciphersuite_name: &CiphersuiteName) -> AeadType {
-    match ciphersuite_name {
-        CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-        | CiphersuiteName::MLS10_128_DHKEMP256_AES128GCM_SHA256_P256 => AeadType::Aes128Gcm,
-        CiphersuiteName::MLS10_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-            AeadType::ChaCha20Poly1305
-        }
-        CiphersuiteName::MLS10_256_DHKEMX448_AES256GCM_SHA512_Ed448
-        | CiphersuiteName::MLS10_256_DHKEMP384_AES256GCM_SHA384_P384
-        | CiphersuiteName::MLS10_256_DHKEMP521_AES256GCM_SHA512_P521 => AeadType::Aes256Gcm,
-        CiphersuiteName::MLS10_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-            AeadType::ChaCha20Poly1305
-        }
-    }
-}
