+++ openmls/src/extensions/capabilities_extension.rs
@@ -93,20 +93,20 @@ impl tls_codec::Deserialize for CapabilitiesExtension {
         }
 
         let ciphersuites = TlsVecU8::<CiphersuiteName>::tls_deserialize(bytes)?;
-        // There must be at least one ciphersuite we support.
-        let mut supported_suite = false;
-        for suite in ciphersuites.iter() {
-            if suite.is_supported() {
-                supported_suite = true;
-                break;
-            }
-        }
-        if !supported_suite {
-            return Err(tls_codec::Error::DecodingError(format!(
-                "{:?}",
-                CapabilitiesExtensionError::UnsupportedCiphersuite,
-            )));
-        }
+        // // There must be at least one ciphersuite we support.
+        // let mut supported_suite = false;
+        // for suite in ciphersuites.iter() {
+        //     if suite.is_supported() {
+        //         supported_suite = true;
+        //         break;
+        //     }
+        // }
+        // if !supported_suite {
+        //     return Err(tls_codec::Error::DecodingError(format!(
+        //         "{:?}",
+        //         CapabilitiesExtensionError::UnsupportedCiphersuite,
+        //     )));
+        // }
 
         let extensions = TlsVecU8::tls_deserialize(bytes)?;
         let proposals = TlsVecU8::tls_deserialize(bytes)?;
