+++ openmls/src/utils.rs
@@ -204,14 +204,3 @@ macro_rules! implement_persistence {
         }
     };
 }
-
-// Implement std::fmt::Display for a given enum type.
-macro_rules! implement_enum_display {
-    ($error:ty) => {
-        impl std::fmt::Display for $error {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                f.write_fmt(format_args!("{:?}", self))
-            }
-        }
-    };
-}
